fix: split feedback modals and Stellar/events components for issue #21504
Split 7 oversized files into sibling sub-components and utilities:

Feedback:
- FeedbackModal.tsx (784→291): extract FeedbackTabBar, ScreenshotAttacher,
  RatingSelector, FeedbackSuccessPanel, FeedbackFormBody, useFeedbackDraft
- SubmitTab.tsx (773→299): extract SubmitFormFields, CategoryPicker,
  SubmitAuthGateBanner, SubmitTokenWarning, SubmitErrorPanel, useSubmitFormHandler
- UpdatesTabRequestItem.tsx (648→297): extract RequestStatusBadge,
  VerificationActions, DiffPreview
- SetupInstructionsDialog.tsx (411→303): extract InstructionStepCard,
  CopyableCommand, VerificationStatus

Stellar/events:
- BatchMonitorModal.tsx (749→291): extract BatchItemRow, BatchSummary,
  BatchProgressBar, BatchMonitorModal.utils
- EventModal.tsx (577→329): extract EventDetailPanel, RawPayloadViewer,
  EventModal.ActionButtons, EventModal.utils, EventInvestigateView
- StellarAuditLogSection.tsx (429→306): extract AuditLogRow, AuditFilterChips,
  AuditExportButton

All originals reduced below 350 lines. No public API or prop changes.
